requirements.txt
streamlit
numpy 
pillow

